+++ e2e-tests/cypress/tests/support/index.js
@@ -28,6 +28,7 @@ import './fetch_commands';
 import './keycloak_commands';
 import './ldap_commands';
 import './ldap_server_commands';
+import './notification_commands';
 import './okta_commands';
 import './saml_commands';
 import './shell';
