+++ e2e-tests/cypress/tests/support/notification_commands.d.ts
@@ -1,0 +1,22 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+/// <reference types="cypress" />
+
+// ***************************************************************
+// Each command should be properly documented using JSDoc.
+// See https://jsdoc.app/index.html for reference.
+// Basic requirements for documentation are the following:
+// - Meaningful description
+// - Each parameter with `@params`
+// - Return value with `@returns`
+// - Example usage with `@example`
+// Custom command should follow naming convention of having `external` prefix, e.g. `externalActivateUser`.
+// ***************************************************************
+
+declare namespace Cypress {
+    interface Chainable {
+        stubNotificationPermission(permission: string): Chainable<void>;
+        verifySystemBotMessageRecieved(): Chainable<void>;
+    }
+}
